Remove anti-override-dependent-values
This mitigation is no longer needed after recent APIv2 changes. The
APIv2 now sends only the one changed value and not all of them.
Therefore we no longer need to do guesswork on what the user intended.

This (now removed) implementation did not work well anyways, it caused
the app to act just upon the first attribute and not the other ones.